+++ miataru/miataru/views/iPhone/Groups Views/iPhone_GroupMapView.swift
@@ -713,17 +713,6 @@ struct iPhone_GroupMapView: View {
                 deviceAccuracies[location.Device] = location.HorizontalAccuracy
                 deviceTimestamps[location.Device] = location.TimestampDate
                 now = Date() // Update time immediately for each hit
-                // Caching: Save new location (including optional speed)
-                DeviceLocationCacheStore.shared.setLocation(
-                    for: location.Device,
-                    latitude: location.Latitude,
-                    longitude: location.Longitude,
-                    accuracy: location.HorizontalAccuracy,
-                    timestamp: location.TimestampDate,
-                    batteryLevel: location.BatteryLevel,
-                    altitude: location.Altitude,
-                    speed: location.Speed
-                )
             }
             // Update map region to fit all devices only when auto-centering is enabled
             if isAutoCenteringEnabled { updateMapRegionToFitDevices() }